Adding utils.go
Also changing fmt to log where possible
