+++ server/Makefile
@@ -13,7 +13,7 @@ GEN_FILE=$(buildpath)/src/gen/librelay/relay_hub_sol.go
 $(server_exe): $(GEN_FILE) $(ETHFILE) $(shell find . -maxdepth 3 -name '*.go') Makefile
 	echo "Using GOPATH=$(GOPATH)"
 	mkdir -p $(buildpath)/bin
-	go build -race -o $(server_exe) src/RelayHttpServer.go
+	go build -race -o $(server_exe) src/RelayHttpServer.go src/utils.go
 
 go-get: $(GEN_FILE) $(ETHFILE)
